refactor: :recycle: extract code
